feat(workspace): Introduce `kona-proof` crate
