+++ Cargo.toml
@@ -70,6 +70,7 @@ kona-executor = { path = "crates/executor", version = "0.0.5", default-features 
 kona-common = { path = "crates/proof-sdk/common", version = "0.0.4", default-features = false }
 kona-common-proc = { path = "crates/proof-sdk/common-proc", version = "0.0.5", default-features = false }
 kona-preimage = { path = "crates/proof-sdk/preimage", version = "0.0.4", default-features = false }
+kona-proof = { path = "crates/proof-sdk/proof", version = "0.0.1", default-features = false }
 
 # Alloy
 alloy-rlp = { version = "0.3.9", default-features = false }
