+++ bin/client/src/handler/kzg_point_eval.rs
@@ -1,34 +1,38 @@
-//! Contains the accelerated version of the `ecrecover` precompile.
+//! Contains the accelerated version of the KZG point evaluation precompile.
 
-use crate::fault::{HINT_WRITER, ORACLE_READER};
+use crate::{HINT_WRITER, ORACLE_READER};
 use alloc::{string::ToString, vec::Vec};
 use alloy_primitives::{keccak256, Address, Bytes};
-use kona_client::{errors::OracleProviderError, HintType};
 use kona_preimage::{
     errors::PreimageOracleError, HintWriterClient, PreimageKey, PreimageKeyType,
     PreimageOracleClient,
 };
+use kona_proof::{errors::OracleProviderError, HintType};
 use revm::{
     precompile::{u64_to_address, Error as PrecompileError, PrecompileWithAddress},
     primitives::{Precompile, PrecompileOutput, PrecompileResult},
 };
 
-const ECRECOVER_ADDRESS: Address = u64_to_address(1);
+const POINT_EVAL_ADDRESS: Address = u64_to_address(0x0A);
 
-pub(crate) const FPVM_ECRECOVER: PrecompileWithAddress =
-    PrecompileWithAddress(ECRECOVER_ADDRESS, Precompile::Standard(fpvm_ecrecover));
+pub(crate) const FPVM_KZG_POINT_EVAL: PrecompileWithAddress =
+    PrecompileWithAddress(POINT_EVAL_ADDRESS, Precompile::Standard(fpvm_kzg_point_eval));
 
-/// Performs an FPVM-accelerated `ecrecover` precompile call.
-fn fpvm_ecrecover(input: &Bytes, gas_limit: u64) -> PrecompileResult {
-    const ECRECOVER_BASE: u64 = 3_000;
+/// Performs an FPVM-accelerated KZG point evaluation precompile call.
+fn fpvm_kzg_point_eval(input: &Bytes, gas_limit: u64) -> PrecompileResult {
+    const GAS_COST: u64 = 50_000;
 
-    if ECRECOVER_BASE > gas_limit {
+    if gas_limit < GAS_COST {
         return Err(PrecompileError::OutOfGas.into());
+    }
+
+    if input.len() != 192 {
+        return Err(PrecompileError::BlobInvalidInputLength.into());
     }
 
     let result_data = kona_common::block_on(async move {
         // Write the hint for the ecrecover precompile run.
-        let hint_data = &[ECRECOVER_ADDRESS.as_ref(), input.as_ref()];
+        let hint_data = &[POINT_EVAL_ADDRESS.as_ref(), input.as_ref()];
         HINT_WRITER
             .write(&HintType::L1Precompile.encode_with(hint_data))
             .await
@@ -63,5 +67,5 @@ fn fpvm_ecrecover(input: &Bytes, gas_limit: u64) -> PrecompileResult {
     })
     .map_err(|e| PrecompileError::Other(e.to_string()))?;
 
-    Ok(PrecompileOutput::new(ECRECOVER_BASE, result_data.into()))
+    Ok(PrecompileOutput::new(GAS_COST, result_data.into()))
 }
