+++ bin/client/src/lib.rs
@@ -6,24 +6,5 @@
 
 extern crate alloc;
 
-pub mod l1;
-
-pub mod l2;
-
-pub mod sync;
-
-pub mod errors;
-
-pub mod executor;
-
-mod hint;
-pub use hint::HintType;
-
-pub mod boot;
-pub use boot::BootInfo;
-
 mod pipe;
 pub use pipe::PipeHandle;
-
-mod caching_oracle;
-pub use caching_oracle::{CachingOracle, FlushableCache};
