+++ crates/proof-sdk/proof/README.md
@@ -1,0 +1,8 @@
+# `kona-proof`
+
+<a href="https://github.com/anton-rs/kona/actions/workflows/rust_ci.yaml"><img src="https://github.com/anton-rs/kona/actions/workflows/rust_ci.yaml/badge.svg?label=ci" alt="CI"></a>
+<a href="https://crates.io/crates/kona-proof"><img src="https://img.shields.io/crates/v/kona-proof.svg?label=kona-proof&labelColor=2a2f35" alt="Kona Proof SDK"></a>
+<a href="https://github.com/anton-rs/kona/blob/main/LICENSE.md"><img src="https://img.shields.io/badge/License-MIT-d1d1f6.svg?label=license&labelColor=2a2f35" alt="License"></a>
+<a href="https://img.shields.io/codecov/c/github/anton-rs/kona"><img src="https://img.shields.io/codecov/c/github/anton-rs/kona" alt="Codecov"></a>
+
+`kona-proof` is an OP Stack state transition proof SDK.
