+++ crates/proof-sdk/proof/src/caching_oracle.rs
