+++ crates/proof-sdk/proof/src/errors.rs
